W3-D4-Parametrrization/requirements.txt
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
joblib==1.4.2
pytest==8.3.2
matplotlib
pyyaml==6.0.2

